Fixed spelling mistake in Flesch Kincaid stubs. Tidied readability code, readability ranking only considers readability once per document. Readability code now warning free on windows.
git-svn-id: https://groke.otago.ac.nz/ant/ant@451 7489a7b4-4588-4892-8acd-774bacfab51c

+++ source/readability_dale_chall.h
@@ -26,9 +26,9 @@ private:
 
 private:
 	ANT_string_pair *measure_name;
-	unsigned long number_of_sentences;
-	unsigned long number_of_words;
-	unsigned long number_of_unfamiliar_words;
+	size_t number_of_sentences;
+	size_t number_of_words;
+	size_t number_of_unfamiliar_words;
 	word *words_encountered;
 	unsigned long last_was_title; // whether the last word we passed on was title
 	unsigned long size;
